ticketing/orders/src/routes/show.ts
import { NotAuthorized, NotFoundError, requireAuth } from "@llticketing/common";
import express, { Request, Response } from "express";
import { Order } from "../models/order";

const router = express.Router();

router.get(
  "/api/orders/:id",
  requireAuth,
  async (req: Request, res: Response) => {
    const order = await Order.findById(req.params.id);

    if (!order) {
      throw new NotFoundError();
    }

    if (order.userId !== req.currentUser!.id) {
      throw new NotAuthorized();
    }

    res.send(order);
  }
);

export { router as showOrderRouter };
